+++ README.md
@@ -2,6 +2,8 @@
 
 ## Usage
 - Run tests: `cargo test --release`
+- Rustfmt: `cargo fmt`
+- Check [typos](https://github.com/crate-ci/typos): `typos -c .github/workflows/typos.toml`
 
 ## Book
 The `book` contains the specification of POD2. A rendered version of the site can be found at: https://0xparc.github.io/pod2/
